Fixed ETCS parser test case.
The problem was as follows:
  a&(b&c) parses to a&(b&c)
  a&(b&c) prints as a&b&c because there's no ambiguity
  But, & is parsed left-associatively so a&b&c parses to (a&b)&c

So the error was caused by the fact that a&(b&c) != (a&b)&c.

For now I fixed this in the parser, but the better way to fix this is to have
a canonical representation for each connective so that equivalence is less
fine-grained.