completedTask.stepId now uses existing IDs
This fixes the but where proofs would not re-load. The problem was that
each step, after completion, had 2 guids -- the one associated with the completed table,
and the id of the original proof. This is somewhat confusing, so now we re-use
the guid as a primary key in the completed table.