Strengthened applicability test for inverse cuts
This might still not apply enough; I'm not entirely sure what
the post-condition of thise phase of the tactic should be. In the
theory, we never have things like "true" and "t >= 0" in the ev
dom constraint, but for the first logical ode solver test we end
up with both. So the question is whether having those things in the
constraint is going to mess up any of the later steps. I think the
answer is no, but if there are things that don't have the correct shape
further down the line, then this is a likely culprit.