Changed/added methods for executing tactics.
These names should cause less confusion.
  BuiltInTactic.apply --> .result, to make it clear that
    BuiltInTactics should still be executed using an interpreter
    instead of directly.

Also added a BelleExpr.execute that runs the tactic with a default
interpreter, for the sake of convienance. This I'm weary about --
using .execute within the implementation of a tactic should be avoided
because the interpeter has a bunch of utility hook-ins that won't work
otherwise.

For example, creating a built-in tactic that calls .execute on another
tactic would create the same problems that we currently experience with
the wub UI when mixing sequent/hilbert styles using the old tactics
framework.