Fixed bug in proof deletion
executionSteps associated with a proof have to be deleted when the proof is deleted
or else a bunch of indices end up wrong somehow.

Or maybe they get loaded even though the proof doesn't exist. Not sure.

At any rate we should fix this in the schema at some point by cascading proof deletions.