Extract tactics without reparsing.
loop({`1=1`}, 1) prints correctly into the database, but is finally extracted as loop(1). I think
this is probably a bug in the pretty printer that is best dealt with by Belle data structure
refactoring.

This commit contains a temporary fix -- instead of printing them parsing I just directly print the
extracted tactic. Note that there are many flaws with this approach, but we can't do much better
until the data structures are refactored anyways.