Fixes bug in parser type analysis
The bug allowed declarations like:

Function.
  R b.
End.

In which case b would be treated like a ProgramVariables.

The bug also allowed for the use of program variables whenever a function of the same name and
sory exists.

The parser now insists that function declarations have a sort, and that all uses of functions are
actually functions (and not program variables of same name.)