Catches errors when prepopulating DerivedLemmasDB.
This allows us to at least try to establish lemmas after the first one fails,
although due to inter-dependencies you still ought to start with the first when
debugging.

Note that I'm using a range instead of a forall or foreach or fold or map because
x => try { ... } already accesses the lazy field, so the error is thrown too early.
And using implicit _ apparently doesn't work because Scala.