Fixes the None.get error message on web UI.
This had multiple causes, all which returned the same error text!

The underlying issue was that the new Lemma mechanism expected we'd never encounter an empty lemma
in the SQLite lemma database, but we did!

There were two independent causes for the Non.get error message:

1) We were writing an empty lemma to the lemma db, which of course doesn't have any Evidence.
   But then when we read the lemma back in, there was no evidence. This is at least ignored due
   to compatability mode for the moment but there might still be an underlying bug here.

2) CachedLemmaDB assumes it's possible to create a lemma with empty contents and then .get that
   lemma. This is used in the creation protocol, which first reserves the name and then updates the
   previous created lemma. But SQLite lemma DB cannot read in an empty lemma because it ends up
   calling .get on a None.

Also, this is many of the None.get errors come from. And it's not quite as simple as always pattern
matching because within Slick code,

blah.getOrElse(x)

is apparently NOT equivalent to

blah match { case Some(y) => y case None => x }

which actualyl doesn't even type check. So we always just call get, which results in None.get
when the field isn't defined.

(I think...)