ODE solver should better support formulas with non-arithmetic postconditions
There is currently no support for formulas whose post-conditions are
non-arithmetic.

I think a somewhat workable solution is to perform all possible equality
rewriting using the cut-in solution on the implication generated in
simplifyPostCondition, and then trying to either QE or else close.

Andre also suggested simply performing box assignments in the post-
condition instead, which should be equivalent in effect.