Added a prop tactic that at least does one level of orR reasoning
There's maybe a language design problem here -- Andre's fixed-point style
tactic should maybe be possible to express somehow?

But in any case I removed Andre's implementation because the reference to "prop"
from within the "prop" tactic results in null pointer exceptions (because "prop"
isn't defined yet).

Note: If you change "prop" to a def then it's impossible to construct
the relevant data structure in the BelleExpr's because prop is just called until
the stack blows up. If we want to support this sort of thing we'll need a language
primitive for it.