Create a new type of Bellerophon error that is always propagated.
Intended for situations where a good error message is written by the built-in tactic author.

This is a first commit toward better error messages.

I propose taking the time to write down an error propagation semantics that takes into acount
things like:
 * Errors in built-ins that are good error messages
 * Errors in built-ins that are possibly bad error messages
 * Errors in built-ins that are known-bad error messages and should be
   replaced with something generic.
 * Traceability

with lots of regression testing and so on.

Also, if a mechanism already exists for doing what I'm using this exception class for (propagating
everything to the top without any wrapping/replacing of the message), please let me know.