Bellerophon: closeTransitive tactic
closeTransitive takes a goal of the form

	a >=b, ..., b>=c|- a >= c

then introduces a specialized transitivity lemma

	\forall abc (a >= b & b >= c -> a >= c)

proves this lemma by a QE call and the properly instantiates the lemma
so that the rest of the proof follows by closeId's (no QE required).
This is nice for subgoals where you want to use transitivity on subgoals
that QE can't handle for whatever reason.