Web UI: work-around for bug in "Hint" portion of prover UI
Whenever a null position is provided, automatically re-write to 1.

TODO: A better fix would be to track down why the int portion of the
webui sometimes passes the wrong positional information to the backend,
and then some up with some systematic guard against this sort of bug.

Also adds a better (or at least unique) error message whenever incorrect
position strings are passed to the web ui.