Upgrade to Scala 2.12
Upgrading to a more recent version of Scala required updating several dependencies.
In most cases, this required changes to .sbt files, but no changes to the code.

The exception, and the bulk of this commit, was switching from Spray to Akka.
I wasn't sure how to do this in small steps, so this is a monolithic commit.