Insist on existence in R during Mathamatica QE.
This is an attempt to fix the fact that Mathematica believes:

    (x^(1/2))^2 = x

is valid and also believes that

    a()^2 >= 0

is valid. Together, these can be used to prove that

    x >= 0

is valid, which is clearly unsound.

Mathematica automatically simplifies  (x^(1/2))^2 = x to True, and I can't find a way to turn
this simplification off. Therefore, I've worked around the problem by insisting that Mathematica also
check that b^e \in R whenever e is not a natural number.

This is a soundness-critical change and possibly a major breaking change. This commit's effect
should be thoroughly investigated and reviewed (or the commit should be reverted)
before a release containing this commit is released to FCPS students.