Adds conversion support for Mathematica's C[].
Mathematica uses C[i] as its default form for parameters/constants.
For example, when computing series approxiamtions to solutions of ODEs, C[i]
is used to represent the constant of integration.

Perhaps a better way of handling this is to ensure that we never get C[i] back from
Mathematica. This can be achieved by always specifying initial conditions in calls to solvers.
For example, instead of:

```
AsymptoticDSolveValue[
  {
    x'[t] == y[t],
    y'[t] == x[t],
    y[0] == x0,
    x[0] == y0
  },
  {x[t], y[t]},
  {t, 0, 10}
]
```

instead do:

```
AsymptoticDSolveValue[
  {
    x'[t] == y[t],
    y'[t] == x[t],
    y[0] == x0,
    x[0] == y0
  },
  {x[t], y[t]},
  {t, 0, 10}
]
```

and then handle x0 and y0 before passing formulas back out of the tool.

Note: this is not currently used in any soundness-critical way that I'm aware of,
but both Reduce and DSolve are capable of generating C[i] symbols. Therefore,
we should do a thorough check that C[i]s are never handed back to us when not expected
or else revert this commit.

See https://reference.wolfram.com/language/ref/C.html for more information on what C[i]
is used for in Mathematica.