Document compilation and dev environment
These files are still missing a lot of info and documentation files are
still all over the place. I plan on adding, moving and updating more doc
files as I learn the code base.

Eventually, all kinds of global documentation should end up as markdown
files in the `doc` directory and be linked to from each other, from the
readme, the wiki, and potentially other places.