Seal subtraits in core
Sealing traits is not transitive, so subtraits need to be sealed
individually. If they are not, they can be implemented from outside the
core, which can be soundness critical. For example, one could implement
a PositionRule that allows proof of anything from nothing.

This also fixes some match exhaustiveness warnings.