Update slick to version 3.2.3
Slick 3.2 is the earliest version after 2.1 that supports Scala 2.12,
which keymaerax currently uses. This means that we'll have to jump
through a few breaking version changes in one go. After that, we can
update to slick 3.4.1 in two further steps.

The next few commits will be fixing all the errors we just caused.