Fix ArrayStoreException when parsing Mathematica results
Caused by weird casting on the part of keymaera and the slightly
different type inference of Scala 2.13.

From what I can tell, what happened is this:

Originally, the type parameter T of convertNaryL is inferred as Term. In
convertNaryL's last line, the subexpressions.map(...) expression would
then return an Array[Term]. At runtime, the JVM stores the type
parameter of arrays (but not other generics) and checks against it when
elements are inserted into the arrays. This is fine because all KExpr
subexpressions were also from Term subclasses.

After updating to Scala 2.13 however, the type parameter T is inferred
as the return type of the argument op, for example Times. To check this,
I obtained the type of T like this:
  val tag: ClassTag[T] = implicitly

Quick aside: When trying to unify a function like
  Times.apply: (Term, Term) => Times
with
  op: (T, T) => T
both T=Term and T=Times are valid solutions because of the variance of
function arguments and return values.

This then leads to the subexpressions.map(...) expression returning an
Array[Times], which is too specific a type in this case. We only know
that the subexpressions are Terms! Because the JVM checks the types of
array contents at runtime, we can't just stick our fingers in our ears,
shout "LALALALA" and use or re-cast the array again afterwards.

My solution is to explicitly annotate the "base" type for our arrays
every time one of these typecast-y convert functions is used, thereby
restoring pre-2.13 behaviour.