Quote string constants used as regex
These constants are expected to be plain strings and usually treated as
such (e.g. used in calls to String.contains). However, they're used
unquoted as argument to String.split, which expects a regex. This may
cause bugs if any of these constants is ever changed.

The solution is quoting the constant whenever it is passed to a function
expecting a regex. That way, it is never interpreted as regex.