Make tests less verbose
Successful tests should not print to stdout. Otherwise, the test output
is cluttered or confusing. It makes ignored and failed tests much harder
to locate.

Only when a test fails, it may produce additional output that helps
locate the issue. However, at that point, you can also modify the test
or attach a debugger to get more info.