Fix interpreter not being killed at the right time
This commit fixes the SequentialInterpreterTests crashing halfway in.

While doing proofs with db.proveBy, the database registers a new
interpreter using BelleInterpreter#setInterpreter with a listener that
uses the database itself. This listener also accesses the database when
the interpreter is killed again.

This interpreter however is only killed in the TacticTestBase#beforeEach
of the next test, after the TacticTestBase#afterEach already closed the
database again. This means that the listener tries to run an action on a
closed database connection, resulting in a RejectedExecutionException
and a crash of SequentialInterpreterTests.

This commit kills (and unsets) the interpreter after each test, and does
so before the db is closed. This fixes the error.