Use Version instead of String
Since we now have a nice Version type, we should use it whenever we
previously handled raw version strings.

Since we don't yet have a consistent way of converting a Version back
into a String, in some cases the version string still needs to be passed
along. In the future, this should be fixed by implementing an
appropriate toString for Version.