Fix KeYmaeraXArchaicArchiveParserTests
This test fails since Enguerrand added some extra checks against
undeclared anonymous dots used as arguments. I changed the error message
a bit and updated the test to use the new error message. I'm not
entirely sure if this is the correct approach for this test, or if the
old error message should somehow be triggered instead.