Remove kaisar cdgl prover
There exists no way to use kaisar to prove anything. It is not reachable
from the CLI or web UI. There is no class with main function anywhere in
the cdgl package. Its code is not used outside of the cdgl package.

The only place where code from cdgl was used is in the pt module. Here,
the ProofCheckException from cdgl was used. All arguments except the
exception message were left on their default values. So I just created a
new ProofTermException that extends Exception and only has a message.

After this single change, the entire cdgl package (and corresponding
tests) could be removed without any additional changes.