Move VersionNumber and BuildInfo to info package
My goal with this package is that it centralizes all project and build
info (as described in the package docstring).

In order to prevent accidentally using BuildInfo directly, it has been
moved and made package private. All project and build info must be
accessed through the `info` package constants instead.