Include third-party licenses in project info
The third-party license info file has been moved next to the normal
LICENSE.txt file from its previous location somewhere deep within the
recesses of the webui resources. This makes it harder to miss.

It has also been included in the project info like the license and
copyright files. In the web ui, it is now displayed verbatim instead of
as a table. This makes the code simpler and more robust against future
changes of the third-party license file.