Fix some CLI options being ignored in ui command
Previously, the CLI argument parser would set some Configuration values
as a side effect. After 7563ec80cc6acfdac979def17db47f792eaf60ff, this
no longer happened, and I forgot to update the Configuration values
manually for the ui command (because it was special-cased in the
launcher instead of the normal CLI).