Fix hashes in ignore-revs file
Due to PRs being rebased during merging and the .git-blame-ignore-revs
file being created in the same PR as the ignored "Reformat everything"
commit, its contents were incorrect. For the same reason, the commit
moving everything to org.keymaerax was also not ignored yet. Now that
everything is on the master branch, the ignore-revs file can be updated.