Fix unreachable code warning
There are two cases for (h: ExpressionArtifact, e: ExpressionArtifact).
Originally, those cases were distinguished by the types in the
ExpressionArtifacts (similar to TexExpressionArtifact), but this type
info was lost in a refactoring, making the two cases identical.

Commit 9fcf9b19cad5cb3289cef534e61818b63b716e1a tries to fix this but
seemingly forgets that the bottom case already has an implementation for
formulas and so introduces code that differs slightly from the original
code for formulas.

In this commit, the original code for formulas is reinstated and the
redundant second case is deleted.