Fix erased type check warnings
Instead of specifying the generics in the match, which leads to compiler
warnings and looks like we're matching on the generic types (which of
course we aren't due to type erasure), the code now explicitly casts.
Aside from fixing a bunch of warnings, this makes it more obvious what's
actually going on here.