Import scala.io classes properly
For some reason, the build started failing once I added the classgraph
library. Classgraph's classes reside in the io.github.classgraph.*
package and subpackages.

My best guess for why this happens is that for backwards compatibility,
io.* resolves to scala.io.* as long as no dependency uses io.*. Maybe
once upon a time, the scala stdlib lived in the root instead of in
scala.*, or maybe there are other (hopefully good) reasons for this.

What made the situation worse is that this did not invalidate the
incremental compilation caches correctly, meaning I only noticed it a
few commits down the line when unrelated files started failing for no
obvious reason.

The obvious and simple fix implemented in this commit is to properly
import or fully specify anything from scala.io.* whenever it is used.