Fix tool path autodetection for Mathematica 14.1
Starting with Mathematica 14.1 from what I can tell, "Mathematica" in
the default installation directory was renamed to "Wolfram". This breaks
tool path autodetection on systems where wolframscript or wolfram are
not on the PATH, i.e. MacOS and possibly Linux.

To fix this, adding the new default installation directories should
suffice since the structure of the installation directory was not
changed, just its path.