Retrieve TacticInfo lazily
Currently, tactics register their own TacticInfo with the DerivationInfo
database when they are constructed. In many cases, this happens whenever
a tactic is applied during a proof as well.

Some (val) tactics are constructed when the object they're defined in is
initialized. Other (lazy val, def) tactics need to be manually
constructed. by evaluating or calling them.

The job of the DerivationInfoRegistry is to manage the DerivationInfo
initialization process. During this process, it rifles through all
objects for tactics and ensure every tactic has been constructed at
least once. It does this using a lot of reflection, and it locates the
tactics using a helper trait called TacticProvider.

This process is very convoluted and fragile. One has to understand lots
of different components (the @Tactic macro implementation, Java and
Scala reflection, the Scala object initialization mechanism, ...) in
order to understand the initialization process.

In the future, the TacticInfo registration will happen in a separate
phase and won't be driven by tactic construction: A tactic's TacticInfo
will be defined as a separate field next to the tactic itself and
annotated with @Derivation. During the registration phase, the registry
will find and store all values annotated with @Derivation.

Since TacticInfo will no longer be globally available immediately after
tactic construction, any val that retrieves TacticInfo now needs to be a
lazy val that is only accessed after the registration phase.

In the future, once the @Tactic macro has been removed, there will no
longer be a need to retrieve TacticInfo since it will be defined
alongside the tactic in a separate field that can be accessed directly.