Enable case-apply-copy-access feature
This feature switches on Scala 3 behavior regarding the visibility of
case class copy and apply methods.

In particular, when the case class constructor is private, the generated
copy and apply methods are private as well.

When the case class constructor is protected, the generated copy method
is protected as well, but the generated apply method in the companion
object is still public (see https://github.com/scala/scala3/pull/5830).

See also https://github.com/scala/scala/pull/7702